% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{chinese_to_arabic}
\alias{chinese_to_arabic}
\title{Converting Chinese into Numbers}
\usage{
chinese_to_arabic(chinese_num)
}
\arguments{
\item{chinese_num}{a Chinese character representing numbers, e.g., "三百二十八".}
}
\value{
converting results, a number.
}
\description{
We provide a chinese_to_arabic function to convert Chinese to Arabic numbers.
This function will be used by the function for getting sentence information.
However, users can also use this function to convert other Chinese numbers.
The maximum unit of converting is one hundred million.
Our converting will consider "两", which is same to "二" in Chinese. The unicode of "两" is u4e24.
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u4e24 \tab "两" \cr
\\u4e8c\\u767e\\u4e09\\u5341\\u516" \tab "二百三十六" \cr
\\u4e24\\u767e\\u4e00\\u5341\\u516b" \tab "两百一十八" \cr
\\u4e94\\u767e\\u516b\\u5341\\u4e09" \tab "五百八十三" \cr
}
}

\examples{
# beginto show results: this function includes "\u4e24".
chinese_to_arabic("\u4e8c\u767e\u4e09\u5341\u516d")
chinese_to_arabic("\u4e24\u767e\u4e00\u5341\u516b")
# you can also provide the variable.
ch_num <- "\u4e94\u767e\u516b\u5341\u4e09"
chinese_to_arabic(ch_num)


}
