% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{extract_defendant_block}
\alias{extract_defendant_block}
\title{Extract Defendant Information Block}
\usage{
extract_defendant_block(text)
}
\arguments{
\item{text}{Full judgment text.}
}
\value{
A data frame with: \code{name} (defendant's name), \code{body} (the sentence without the name), \code{raw} (the full sentence), \code{judgement_year} (the year this judgement is produced).
}
\description{
This function is to extract the certain defendant information block (a long sentence) from the full text.
As we have mentioned, you can use this function combined with your own rule to extract certain features that we do not include.
}
