% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{extract_defendant_extra}
\alias{extract_defendant_extra}
\title{Extract Defendant Extra-legal Information}
\usage{
extract_defendant_extra(text, first_only = TRUE)
}
\arguments{
\item{text}{Full judgment text.}

\item{first_only}{Logical; if TRUE, only the first defendant (the primary defendant) is processed. The default value is \code{TRUE}.}
}
\value{
A data frame with the 11 extra-legal variables. Including name, gender, born-year, born-date, age, ethnic info., education, hukou, hometown, occupation, hukou type
}
\description{
This function is to extract several common extra-legal information from the full document.
The logic behind this is: we first extract the certain block (a long sentence) from the full text.
Next, we extract all concerned features through designed regular expression.
If you need to extract other information, you can use \code{extract_defendant_block} function to get the certain block, and then use your own rule to extract certain features.
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u516c...\\u3002 \tab "公诉机关广东省广州市人民检察院。被告人童桂连，女，1990年6月2日出生，汉族，文化程度高中，无业，高中学历，户籍所在地广东省云浮市云安区，住云浮市云城区。因本案于2019年3月2日被刑事拘留，同年4月4日被逮捕。2021年2月10日被取保候审。被告人袁林林，女，1992年出生。辩护人卢翠玲，广东中泽律师事务所律师。2019年2月，为牟取非法利益，被告人童桂连从广州前往荷兰王国帮助他人运送行李箱至泰王国。" \cr
}
}

\section{Some Variables' Values}{

\tabular{ll}{
  Variable \tab Corresponding Value \cr
  Ethnic Info. \tab ethnic groups, e.g., Han, Zhuang, Manchu, ..., totally 56 values) \cr
  Education \tab education level categories (postgraduate, bachelor, associate, high school, middle school, primary, illiterate) \cr
  Occupation \tab unemployed, farmer, self-employed, worker, student, teacher, doctor, ..., totally 33 values \cr
  Hukou Type \tab agricultural, non-agricultural, urban, rural)\cr
}
}

\examples{
text <- "\u516c\u8bc9\u673a\u5173\u5e7f\u4e1c\u7701\u5e7f\u5dde\u5e02\u4eba\u6c11\u68c0\u5bdf\u9662\u3002\u88ab\u544a\u4eba\u7ae5\u6842\u8fde\uff0c\u5973\uff0c1990\u5e746\u67082\u65e5\u51fa\u751f\uff0c\u6c49\u65cf\uff0c\u6587\u5316\u7a0b\u5ea6\u9ad8\u4e2d\uff0c\u65e0\u4e1a\uff0c\u9ad8\u4e2d\u5b66\u5386\uff0c\u6237\u7c4d\u6240\u5728\u5730\u5e7f\u4e1c\u7701\u4e91\u6d6e\u5e02\u4e91\u5b89\u533a\uff0c\u4f4f\u4e91\u6d6e\u5e02\u4e91\u57ce\u533a\u3002\u56e0\u672c\u6848\u4e8e2019\u5e743\u67082\u65e5\u88ab\u5211\u4e8b\u62d8\u7559\uff0c\u540c\u5e744\u67084\u65e5\u88ab\u902e\u6355\u30022021\u5e742\u670810\u65e5\u88ab\u53d6\u4fdd\u5019\u5ba1\u3002\u88ab\u544a\u4eba\u8881\u6797\u6797\uff0c\u5973\uff0c1992\u5e74\u51fa\u751f\u3002\u8fa9\u62a4\u4eba\u5362\u7fe0\u73b2\uff0c\u5e7f\u4e1c\u4e2d\u6cfd\u5f8b\u5e08\u4e8b\u52a1\u6240\u5f8b\u5e08\u30022019\u5e742\u6708\uff0c\u4e3a\u725f\u53d6\u975e\u6cd5\u5229\u76ca\uff0c\u88ab\u544a\u4eba\u7ae5\u6842\u8fde\u4ece\u5e7f\u5dde\u524d\u5f80\u8377\u5170\u738b\u56fd\u5e2e\u52a9\u4ed6\u4eba\u8fd0\u9001\u884c\u674e\u7bb1\u81f3\u6cf0\u738b\u56fd\u3002"
# consider both defendants
extract_defendant_extra(text, FALSE)
# consider only the primary defendant (the first one)
extract_defendant_extra(text, TRUE)
# if you want to extract all from a df, run:
# result_list <- lapply(df[[text_col]], function(txt) {
#   if (is.na(txt) || txt == "") return(NULL)
#   extract_defendant_extra(txt)
# })

}
