% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{extract_judge_name}
\alias{extract_judge_name}
\title{Extract Judge Name Function}
\usage{
extract_judge_name(text)
}
\arguments{
\item{text}{a full text of judicial documents}
}
\value{
the name of judge
}
\description{
This function can extract the main judge's name.
We will ignore other associate judges' names like other research has done.
We use four regular expressions to extract the name of judges.
It considers both situations when there is only one judge and there are over one judges.
If you need deal with a data.frame's texts, you may consider use \code{for} or \code{lapply} function.
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u5ba1...\\u4e09 \tab "审 判 长 章 三 " \cr
\\u7531...\\u7406 \tab "由审判员欧阳明独任审理" \cr
\\u5ba1...\\u5de5 \tab "审判长:吴工" \cr
}
}

\examples{
df <- data.frame(document = c("\u5ba1 \u5224 \u957f \u7ae0 \u4e09","\u7531\u5ba1\u5224\u5458\u6b27\u9633\u660e\u72ec\u4efb\u5ba1\u7406", "\u5ba1\u5224\u957f:\u5434\u5de5"))
for (i in 1:length(df$document)){
  df$name[i] <- extract_judge_name(df$document[i])
  }
# print the name
print(df$name)

}
