% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{get_dtm_tfidf}
\alias{get_dtm_tfidf}
\title{Get Document Term Matrix With Tf-Idf Weight}
\usage{
get_dtm_tfidf(text)
}
\arguments{
\item{text}{a long text, better to be a column of texts from a data frame.}
}
\value{
a matrix, each row contains a text's vector information.
}
\description{
We provide a get_dtm_tfidf function to get the DTM of a list of texts for further analysis.
We make use of tm package and chinese.misc package to calculate this matrix. To cut sentences into tokens, we use jiebaR.
This function contains a tf-idf weight function to generate a matrix.
From Wiki: tf-idf is a measure of importance of a word to a document in a collection or corpus,
adjusted for the fact that some words appear more frequently in general
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u6211\\u662f\\u4e00...\\u6cd5\\u5f8b\\u5e08 \tab "我是一个刑事诉讼法律师" \cr
\\u6211\\u662f\\u4e00\\u4e2a\\u5f8b\\u5e08 \tab "我是一个律师" \cr
}
}

\examples{
# please rememver that, we need a column from a data.frame.
df <- data.frame(text = c("\u6211\u662f\u4e00\u4e2a\u5211\u4e8b\u8bc9\u8bbc\u6cd5\u5f8b\u5e08",
                 "\u6211\u662f\u4e00\u4e2a\u5f8b\u5e08"),
                  stringsAsFactors = FALSE
                 )
# this function only requires the "text" column.
get_dtm_tfidf(df$text)

}
