% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{get_legal_factor}
\alias{get_legal_factor}
\title{Legal Information Extraction Function}
\usage{
get_legal_factor(maindoc, legaldoc, keyword, legal_article)
}
\arguments{
\item{maindoc}{the reasoning text}

\item{legaldoc}{the legal article text from the appendix}

\item{keyword}{the legal factor you want to get, e.g. "认罪认罚"}

\item{legal_article}{the legal factor's corresponding legal article number, e.g., "第三十条"}
}
\value{
whether there is a legal factor that users care about. yes = 1, no = 0.
}
\description{
We provide a get_legal_factor function that can be used to extract legal factors users want.
The method is "regular expression", and we have already provide a relatively stable, cheap method.
We require users to write the keywords, and also provide the legal_article numbers in Chinese.
Then, users should also input the reasoning text and the legal article part of the sentencing document.
Only when keywords and legal_article numbers can be found in reasoning text and legal article text, we return a value 1 representing there is a factor here.
Otherwise, the function will return a 0.
Please note that, in this function, we can not extract the amount of property crimes. To extract the amount, one needs to ask help from LLMs or design their own regular expressions.
However, this function can save much time and also be used as a baseline function.
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u672c\\u9662...\\u8ba4\\u7f6a...\\u5211\\u7f5a \tab "本院认为，被告人犯寻衅滋事罪，自首，且认罪认罚，因此可以酌情减免刑罚" \cr
\\u672c\\u9662...\\u5211\\u7f5a \tab "本院认为，被告人犯寻衅滋事罪，自首，因此可以酌情减免刑罚" \cr
\\u7b2c\\u4e8c...\\u7406\\u3002 \tab 此段为刑法第二百九十三条与刑事诉讼法第十五条 \cr
\\u7b2c\\u5341\\u4e94\\u6761 \tab "第十五条" \cr
\\u7b2c\\u4e8c\\u767e\\u4e5d\\u5341\\u4e09\\u6761 \tab "第二百九十三条" \cr
\\u8ba4\\u7f6a\\u8ba4\\u7f5a \tab "认罪认罚" \cr
\\u5177\\u7ed3\\u4e66 \tab "具结书" \cr
}
}

\examples{
# note: please get reason and appendix use function we provide. this is for illustration.
reason <- data.frame(text = c("\u672c\u9662\u8ba4\u4e3a\uff0c\u88ab\u544a\u4eba\u72af\u5bfb\u8845\u6ecb\u4e8b\u7f6a\uff0c\u81ea\u9996\uff0c\u4e14\u8ba4\u7f6a\u8ba4\u7f5a\uff0c\u56e0\u6b64\u53ef\u4ee5\u914c\u60c5\u51cf\u514d\u5211\u7f5a", "\u672c\u9662\u8ba4\u4e3a\uff0c\u88ab\u544a\u4eba\u72af\u5bfb\u8845\u6ecb\u4e8b\u7f6a\uff0c\u81ea\u9996\uff0c\u56e0\u6b64\u53ef\u4ee5\u914c\u60c5\u51cf\u514d\u5211\u7f5a"))
appendix <- data.frame(text = c("\u7b2c\u4e8c\u767e\u4e5d\u5341\u4e09\u6761\u6709\u4e0b\u5217\u5bfb\u8845\u6ecb\u4e8b\u884c\u4e3a\u4e4b\u4e00\uff0c\u7834\u574f\u793e\u4f1a\u79e9\u5e8f\u7684\uff0c\u5904\u4e94\u5e74\u4ee5\u4e0b\u6709\u671f\u5f92\u5211\u3001\u62d8\u5f79\u6216\u8005\u7ba1\u5236\uff1a\u5211\u4e8b\u8bc9\u8bbc\u6cd5\u7b2c\u5341\u4e94\u6761\uff1a\u7b2c\u5341\u4e94\u6761\u3000\u72af\u7f6a\u5acc\u7591\u4eba\u3001\u88ab\u544a\u4eba\u81ea\u613f\u5982\u5b9e\u4f9b\u8ff0\u81ea\u5df1\u7684\u7f6a\u884c\uff0c\u627f\u8ba4\u6307\u63a7\u7684\u72af\u7f6a\u4e8b\u5b9e\uff0c\u613f\u610f\u63a5\u53d7\u5904\u7f5a\u7684\uff0c\u53ef\u4ee5\u4f9d\u6cd5\u4ece\u5bbd\u5904\u7406\u3002", "\u7b2c\u4e8c\u767e\u4e5d\u5341\u4e09\u6761\u6709\u4e0b\u5217\u5bfb\u8845\u6ecb\u4e8b\u884c\u4e3a\u4e4b\u4e00\uff0c\u7834\u574f\u793e\u4f1a\u79e9\u5e8f\u7684\uff0c\u5904\u4e94\u5e74\u4ee5\u4e0b\u6709\u671f\u5f92\u5211\u3001\u62d8\u5f79\u6216\u8005\u7ba1\u5236\uff1a\u5211\u4e8b\u8bc9\u8bbc\u6cd5\u7b2c\u5341\u4e94\u6761\uff1a\u7b2c\u5341\u4e94\u6761\u3000\u72af\u7f6a\u5acc\u7591\u4eba\u3001\u88ab\u544a\u4eba\u81ea\u613f\u5982\u5b9e\u4f9b\u8ff0\u81ea\u5df1\u7684\u7f6a\u884c\uff0c\u627f\u8ba4\u6307\u63a7\u7684\u72af\u7f6a\u4e8b\u5b9e\uff0c\u613f\u610f\u63a5\u53d7\u5904\u7f5a\u7684\uff0c\u53ef\u4ee5\u4f9d\u6cd5\u4ece\u5bbd\u5904\u7406\u3002"))
# I want to find whether the defendant plead guilty, corresponding article number 15.
target <- "\u8ba4\u7f6a\u8ba4\u7f5a"
articleNum <- "\u7b2c\u5341\u4e94\u6761"
## 01. run the function: one keyword, one article.
get_legal_factor(maindoc = reason$text, legaldoc = appendix$text, keyword = target, legal_article = articleNum)
# you may want to have several corresponding articles referring to single factor
articleNum <- c("\u7b2c\u5341\u4e94\u6761", "\u7b2c\u4e8c\u767e\u4e5d\u5341\u4e09\u6761")
## 02: run the function: one keyword, two articles.
get_legal_factor(maindoc = reason$text, legaldoc = appendix$text, keyword = target, legal_article = articleNum)
# or you want to have several keywords to make sure your extraction is correct.
target <- c("\u8ba4\u7f6a\u8ba4\u7f5a", "\u5177\u7ed3\u4e66")
## 03: run the function: two keywords, two articles.
get_legal_factor(maindoc = reason$text, legaldoc = appendix$text, keyword = target, legal_article = articleNum)

}
