% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{get_prison_month}
\alias{get_prison_month}
\title{Get Sentencing Data in Months}
\usage{
get_prison_month(text)
}
\arguments{
\item{text}{a long text containing sentencing information}
}
\value{
the length of sentencing in months of imprisonment
}
\description{
We provide a get_prison_month function to extract the sentencing results from one single sentencing document.
This function only considers "有期徒刑", imprisonment.
For "无期徒刑", life sentence, we will return an Inf value.
Other two kinds of sentencing, "拘役" and "管制" -- criminal detention and public surveillance, will be ignored, like many other studies have done.
The unit of output is "month".
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u88ab\\u544a\\u4eba...\\u4e2a\\u6708 \tab "被告人王某犯寻衅滋事罪，判处有期徒刑两年六个月" \cr
}
}

\examples{
# directly convert the sentence.
get_prison_month("\u88ab\u544a\u4eba\u738b\u67d0\u72af\u5bfb\u8845\u6ecb\u4e8b\u7f6a\uff0c\u5224\u5904\u6709\u671f\u5f92\u5211\u4e24\u5e74\u516d\u4e2a\u6708")
# you can also provide a variable.
sentence_text <- "\u88ab\u544a\u4eba\u738b\u67d0\u72af\u5bfb\u8845\u6ecb\u4e8b\u7f6a\uff0c\u5224\u5904\u6709\u671f\u5f92\u5211\u4e24\u5e74\u516d\u4e2a\u6708"
get_prison_month(sentence_text)


}
