% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{get_probation}
\alias{get_probation}
\title{Get Probation Data in Month}
\usage{
get_probation(text)
}
\arguments{
\item{text}{a long text containing sentencing information}
}
\value{
two values: whether sentenced to probation, yes = 1, no = 0; the length of probation in month.
}
\description{
We provide a get_probation function to extract the "缓刑", probation, results from one single sentencing document.
This function only considers probation and its lenght.
The unit of the probation length output is "month".
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u88ab\\u544a\\u4eba...\\u5211\\u4e24\\u5e74 \tab "被告人王某犯寻衅滋事罪，判处有期徒刑六个月, 缓刑两年" \cr
\\u88ab\\u544a\\u4eba...\\u4e09\\u4e2a\\u6708 \tab "被告人王某犯盗窃罪，判处有期徒刑八个月, 缓刑三个月" \cr
}
}

\examples{
# directly convert the sentence.
get_probation("\u88ab\u544a\u4eba\u738b\u67d0\u72af\u5bfb\u8845\u6ecb\u4e8b\u7f6a\uff0c\u5224\u5904\u6709\u671f\u5f92\u5211\u516d\u4e2a\u6708\uff0c\u7f13\u5211\u4e24\u5e74")
# you can also provide a variable.
probation_text <- "\u88ab\u544a\u4eba\u738b\u67d0\u72af\u76d7\u7a83\u7f6a\uff0c\u5224\u5904\u6709\u671f\u5f92\u5211\u516b\u4e2a\u6708,\uff0c\u7f13\u5211\u4e09\u4e2a\u6708"
get_probation(probation_text)

}
