% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{get_sentence_day}
\alias{get_sentence_day}
\title{Get Sentencing Data in Days}
\usage{
get_sentence_day(text)
}
\arguments{
\item{text}{a long text containing sentencing information}
}
\value{
the length of sentencing in days of imprisonment
}
\description{
We provide a get_sentence_day function to extract the sentencing results from one single sentencing document.
This function considers "有期徒刑", "无期徒刑" and "拘役" and "管制" -- imprisonment, life sentence, criminal detention, public surveillance.
Again, "无期徒刑", life sentence, will return an Inf.
The unit of output is "days". One year is considered as 365 days, and one month is considered as 30 days.
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u88ab\\u544a\\u4eba...\\u516b\\u5929 \tab "被告人王某犯寻衅滋事罪，判处拘役六个月八天" \cr
\\u88ab\\u544a\\u4eba...\\u5236\\u4e5d\\u5929 \tab "被告人王某犯寻衅滋事罪，判处管制九天" \cr
}
}

\examples{
# directly convert the sentence.
get_sentence_day("\u88ab\u544a\u4eba\u738b\u67d0\u72af\u5bfb\u8845\u6ecb\u4e8b\u7f6a\uff0c\u5224\u5904\u62d8\u5f79\u516d\u4e2a\u6708\u516b\u5929")
# you can also provide a variable.
sentence_text <- "\u88ab\u544a\u4eba\u738b\u67d0\u72af\u5bfb\u8845\u6ecb\u4e8b\u7f6a\uff0c\u5224\u5904\u62d8\u5f79\u516d\u4e2a\u6708\u516b\u5929"
get_sentence_day(sentence_text)

}
