% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{homo_index}
\alias{homo_index}
\title{No-Reasoning Function}
\usage{
homo_index(matrix, output = c("mean", "original"))
}
\arguments{
\item{matrix}{the DTM matrix you get using get_dtm function or other methods you employs.}

\item{output}{can be "mean" or "original". "mean" will provide the index, while "original" provide the similarity data.}
}
\value{
an index, or a set of similarities.
}
\description{
We provide a homo_index function that can be used to analyze Chinese judges' no-reasoning problems.
Compared with other regions, Chinese sentencing documents are criticized that they do not provide enough reasoning.
We first sum up all cases and get a vector as the baseline, and use cosine similarity to calculate the similarity between each case and this baseline to get a series of data, and get the mean of these data.
The mean is the index. We also provide original data, which is the series of data, because users may want to give a weight to them.
The logic behind this is the overall similarity between reasoning texts.
The detail of this logic is that, if most judges do not provide enough reasoning, it means that they simply copy from others' sayings. Thus, we can use this index to reprent it.
This functions can be used to calculate a series of data.
homo_index uses cosine similarity because it is not sensitive to length.
}
\examples{
# this function requires a matrix, we create a 3 rwo, 4 column matrix
mat <- matrix(c(1,0,0,0,1,1,0,0,1,1,1,0), nrow = 3, ncol = 4)
# show data structure
print(mat)
# you can get index, or get original data
homo_index(matrix = mat, output = "mean")
# not that this function will ignore NAs automatically.
homo_index(matrix = mat, output = "original")

}
