% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{ngenderR_CN}
\alias{ngenderR_CN}
\title{Guess Gender Function}
\usage{
ngenderR_CN(judgename)
}
\arguments{
\item{judgename}{judge's name in Chines, should be a full name. It should be a character variable.}
}
\value{
male = 1, female = 0; and the probability of being that gender.
}
\description{
We provide a ngenderR_CN function that can be used to guess the gender of a name.
The method is originally a python package and we transfer this into a R function.
The mathematical reason behind the method can be found here: https://github.com/observerss/ngender.
This function is to get judges' gender information through their names.
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u8d75\\u672c\\u5c71 \tab "赵本山" \cr
\\u5b8b\\u4e39\\u4e39 \tab "宋丹丹" \cr
}
}

\examples{
# male example
name1 <- "\u8d75\u672c\u5c71"
ngenderR_CN(name1)
# female example
name2 <- "\u5b8b\u4e39\u4e39"
ngenderR_CN(name2)

}
