% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{sim_correlation}
\alias{sim_correlation}
\title{Similarity Function -- Correlation Similarity}
\usage{
sim_correlation(a, b)
}
\arguments{
\item{a}{the baseline}

\item{b}{a series of data being compared with the baseline}
}
\value{
a series of num. data
}
\description{
These functions can be used to calculate a series of data,
while functions in other packages can not do it.
These functions are very basic and we also allow users to make use of them.
Usually we consider the input is a matrix. However, you can try other as well.
However, other functions will make use of these functions in a correct way.
So, it is not a must to know the detail of the function. If you are interested, please go to the package's original codes.
}
