% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{split_judgment_df}
\alias{split_judgment_df}
\title{Cutting Function - DataFrame}
\usage{
split_judgment_df(df, text_col = "document")
}
\arguments{
\item{df}{a data frame. We recommend you to re-name the document column as \code{document}.}

\item{text_col}{the name of the column you want to cut. Default name is \code{"document"}. Put a character here.}
}
\value{
three new columns added to the original \code{df}.
}
\description{
This function is for cutting a set of documents.
It is an extension to Cutting Function - Single Case.
You can also use \code{for} or \code{lapply} and Cutting Function - Single Case (\code{split_judgment}) to replace this function.
}
\section{Unicode Meaning in Examples}{

\tabular{ll}{
 Unicode \tab Meaning \cr
\\u7ecf\\u5ba1...\\u5f92\\u5211 \tab "经审理查明：盗窃行为。本院认为是犯罪。判决如下:五年有期徒刑" \cr
\\u7ecf\\u5ba1...\\u65e0\\u7f6a \tab "经审理查明：抢劫行为。本院认为不构成犯罪。判决如下:无罪" \cr
}
}

\examples{
# build the dataframe
df <- data.frame(document = c("\u7ecf\u5ba1\u7406\u67e5\u660e\uff1a\u76d7\u7a83\u884c\u4e3a\u3002\u672c\u9662\u8ba4\u4e3a\u662f\u72af\u7f6a\u3002\u5224\u51b3\u5982\u4e0b:\u4e94\u5e74\u6709\u671f\u5f92\u5211", "\u7ecf\u5ba1\u7406\u67e5\u660e\uff1a\u62a2\u52ab\u884c\u4e3a\u3002\u672c\u9662\u8ba4\u4e3a\u4e0d\u6784\u6210\u72af\u7f6a\u3002\u5224\u51b3\u5982\u4e0b:\u65e0\u7f6a"))
result <- split_judgment_df(df, "document")
# print result
print(result)

}
