% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legal_text_stats.R
\name{TLCA}
\alias{TLCA}
\title{Treating Like Cases Alike Function}
\usage{
TLCA(matrix, method = c("cosine", "jaccard", "correlation", "euclidean"), k)
}
\arguments{
\item{matrix}{the DTM matrix you get using get_dtm function or other methods you employs.}

\item{method}{can be "cosine", "jaccard", "correlation" or "euclidean", which leads to a usage of corresponding similarity calculation method.}

\item{k}{use the first k cases as the baseline cases. The default value is total number of your case in the matrix.}
}
\value{
a matrix contains similarity data. Each data represent the similarity between the row and column case.
}
\description{
We provide a TLCA function that can complete the analysis by Chen et. al.'s paper on Law probability and risk in a single row of data.
This functions can be used to calculate a series of data,
The detail of this method should be found on https://doi.org/10.1093/lpr/mgag007
The original method of this paper is RWMD, but it only affects the DTM.
Users are welcome to provide their own DTM using LLMs, RWMD, Bert or other methods.
}
\examples{
# assuming we have a matrix contains 3 rows and 4 columns, which means 3 cases with 4 factors
mat <- matrix(c(1,0,0,0,1,1,0,0,1,1,1,0), nrow = 3, ncol = 4)
# show data structure
print(mat)
# the default value of \code{k} is 3 in this example.
TLCA(mat, method = "cosine")

}
